+++ src/index.d.ts
@@ -17,6 +17,7 @@ import {
 
 import {
 	globalConfig,
+	initializeEnvVar,
 	initializeService
 } from './services/config.js';
 
@@ -113,6 +114,7 @@ import {
 	contentStatusCompletedMany,
 	contentStatusReset,
 	contentStatusStarted,
+	extractFromRecordId,
 	flushWatchSession,
 	generateRecordId,
 	getAllCompleted,
@@ -210,6 +212,14 @@ import {
 import {
 	createTestUser
 } from './services/liveTesting.ts';
+
+import {
+	acceptInvite,
+	createAccount,
+	createInvites,
+	fetchUsersMultiAccountDetails,
+	rescindInvite
+} from './services/multi-user-accounts/multi-user-accounts.ts';
 
 import {
 	emitProgressSaved,
@@ -454,6 +464,7 @@ import {
 
 declare module 'musora-content-services' {
 	export {
+		acceptInvite,
 		addContextToContent,
 		addContextToLearningPaths,
 		addItemToPlaylist,
@@ -476,8 +487,10 @@ declare module 'musora-content-services' {
 		contentStatusReset,
 		contentStatusStarted,
 		convertToTimeZone,
+		createAccount,
 		createComment,
 		createForumCategory,
+		createInvites,
 		createPlaylist,
 		createPost,
 		createPracticeNotes,
@@ -500,6 +513,7 @@ declare module 'musora-content-services' {
 		editComment,
 		emitProgressSaved,
 		enrollUserInGuidedCourse,
+		extractFromRecordId,
 		extractSanityUrl,
 		fetchAll,
 		fetchAllFilterOptions,
@@ -594,6 +608,7 @@ declare module 'musora-content-services' {
 		fetchUserPracticeMeta,
 		fetchUserPracticeNotes,
 		fetchUserPractices,
+		fetchUsersMultiAccountDetails,
 		filterCoursesInCourseCollections,
 		findIncompleteLesson,
 		flushWatchSession,
@@ -667,6 +682,7 @@ declare module 'musora-content-services' {
 		globalConfig,
 		guidedCourses,
 		hasAnyMethodV2IntroCompleted,
+		initializeEnvVar,
 		initializeService,
 		isBucketUrl,
 		isContentLiked,
@@ -717,6 +733,7 @@ declare module 'musora-content-services' {
 		reportComment,
 		reportPlaylist,
 		requestEmailChange,
+		rescindInvite,
 		reset,
 		resetAllAwards,
 		resetAllLearningPaths,