+++ src/index.js
@@ -21,6 +21,7 @@ import {
 
 import {
 	globalConfig,
+	initializeEnvVar,
 	initializeService
 } from './services/config.js';
 
@@ -117,6 +118,7 @@ import {
 	contentStatusCompletedMany,
 	contentStatusReset,
 	contentStatusStarted,
+	extractFromRecordId,
 	flushWatchSession,
 	generateRecordId,
 	getAllCompleted,
@@ -214,6 +216,14 @@ import {
 import {
 	createTestUser
 } from './services/liveTesting.ts';
+
+import {
+	acceptInvite,
+	createAccount,
+	createInvites,
+	fetchUsersMultiAccountDetails,
+	rescindInvite
+} from './services/multi-user-accounts/multi-user-accounts.ts';
 
 import {
 	emitProgressSaved,
@@ -453,6 +463,7 @@ import {
 } from './services/userActivity.js';
 
 export {
+	acceptInvite,
 	addContextToContent,
 	addContextToLearningPaths,
 	addItemToPlaylist,
@@ -475,8 +486,10 @@ export {
 	contentStatusReset,
 	contentStatusStarted,
 	convertToTimeZone,
+	createAccount,
 	createComment,
 	createForumCategory,
+	createInvites,
 	createPlaylist,
 	createPost,
 	createPracticeNotes,
@@ -499,6 +512,7 @@ export {
 	editComment,
 	emitProgressSaved,
 	enrollUserInGuidedCourse,
+	extractFromRecordId,
 	extractSanityUrl,
 	fetchAll,
 	fetchAllFilterOptions,
@@ -593,6 +607,7 @@ export {
 	fetchUserPracticeMeta,
 	fetchUserPracticeNotes,
 	fetchUserPractices,
+	fetchUsersMultiAccountDetails,
 	filterCoursesInCourseCollections,
 	findIncompleteLesson,
 	flushWatchSession,
@@ -666,6 +681,7 @@ export {
 	globalConfig,
 	guidedCourses,
 	hasAnyMethodV2IntroCompleted,
+	initializeEnvVar,
 	initializeService,
 	isBucketUrl,
 	isContentLiked,
@@ -716,6 +732,7 @@ export {
 	reportComment,
 	reportPlaylist,
 	requestEmailChange,
+	rescindInvite,
 	reset,
 	resetAllAwards,
 	resetAllLearningPaths,