+++ src/services/progress-row/rows/content-card.js
@@ -19,31 +19,32 @@ import { PARENT_ID_TOP_LEVEL } from '../../sync/models/ContentProgress'
  * Fetch any content IDs with some progress, include the userPinnedItem,
  * and generate a map of the cards keyed by the content IDs
  */
-export async function getContentCardMap(brand, limit, userPinnedItem ){
+export async function getContentCardMap(brand, limit, userPinnedItem) {
   const metadata = {
-      brand: brand,
-      contentTypes: Object.values(recentTypes.homeRow),
-      parentId: PARENT_ID_TOP_LEVEL,
-    }
+    brand: brand,
+    contentTypes: Object.values(recentTypes.homeRow),
+    parentId: PARENT_ID_TOP_LEVEL,
+  }
   let recentContentIds = await getAllStartedOrCompleted({ metadata, limit })
   if (userPinnedItem?.progressType === 'content') {
     recentContentIds.push(userPinnedItem.id)
   }
 
-  let contents = recentContentIds.length > 0
-    ? await addContextToContent(
-      fetchByRailContentIds,
-      recentContentIds,
-      'progress-tracker',
-      brand,
-      {
-        addNavigateTo: true,
-        addProgressStatus: true,
-        addProgressPercentage: true,
-        addProgressTimestamp: true,
-      }
-    )
-    : []
+  let contents =
+    recentContentIds.length > 0
+      ? await addContextToContent(
+          fetchByRailContentIds,
+          recentContentIds,
+          'progress-tracker',
+          brand,
+          {
+            addNavigateTo: true,
+            addProgressStatus: true,
+            addProgressPercentage: true,
+            addProgressTimestamp: true,
+          }
+        )
+      : []
   contents = postProcessBadge(contents)
 
   const contentCards = await Promise.all(generateContentPromises(contents))
@@ -62,7 +63,8 @@ function generateContentPromises(contents) {
   contents.forEach((content) => {
     const type = content.type
     if (!allRecentTypeSet.has(type)) return
-    let childHasParent = Array.isArray(content.parent_content_data) && content.parent_content_data.length > 0
+    let childHasParent =
+      Array.isArray(content.parent_content_data) && content.parent_content_data.length > 0
     if (!childHasParent) {
       promises.push(processContentItem(content))
       if (showsLessonTypes.includes(type)) {
@@ -80,7 +82,7 @@ export async function processContentItem(content) {
   const isLive = content.isLive ?? false
   let ctaText = getDefaultCTATextForContent(content, contentType)
 
-  const {completedChildren, allChildren} = await getCompletedChildren(content, contentType)
+  const { completedChildren, allChildren } = await getCompletedChildren(content, contentType)
   content.completed_children = completedChildren
   content.all_children = allChildren
 
@@ -125,7 +127,7 @@ export async function processContentItem(content) {
       isLocked: content.is_locked ?? false,
       subtitle:
         collectionLessonTypes.includes(content.type) || content.lesson_count > 1
-          ? `${content.completed_children} of ${content.all_children ?? content.lesson_count ?? content.child_count} Lessons Complete`
+          ? `${content.completed_children ?? 0} of ${content.all_children ?? content.lesson_count ?? content.child_count} Lessons Complete`
           : (contentType === 'lesson' || contentType === 'show') && isLive === false
             ? `${content.progressPercentage}% Complete`
             : `${content.difficulty_string} • ${content.artist_name}`,
@@ -163,8 +165,9 @@ function getDefaultCTATextForContent(content, contentType) {
 }
 
 async function getCompletedChildren(content, contentType) {
-  let completedChildren = null
-  let allChildren = null
+  let completedChildren = 0
+  let allChildren = 0
+
   if (contentType === 'show') {
     const shows = await addContextToContent(fetchShows, content.brand, content.type, {
       addProgressStatus: true,
@@ -173,7 +176,7 @@ async function getCompletedChildren(content, contentType) {
       (show) => show.progressStatus === 'completed'
     ).length
     allChildren = Object.values(shows).length
-  } else if (content.lesson_count > 0) {
+  } else if (content.children && content.children.length > 0) {
     const lessonIds = getLeafNodes(content)
     const progressOnItems = await getProgressStateByIds(lessonIds)
     completedChildren = Array.from(progressOnItems.values()).filter(
@@ -181,7 +184,8 @@ async function getCompletedChildren(content, contentType) {
     ).length
     allChildren = lessonIds.length
   }
-  return {completedChildren, allChildren}
+
+  return { completedChildren, allChildren }
 }
 
 function getLeafNodes(content) {